publish jdcloud-sdk-go version 1.0.7

+++ services/waf/models/LogReportReq.go
@@ -29,5 +29,5 @@ type LogReportReq struct {
     LogId string `json:"logId"`
 
     /* 产生时间  */
-    AccessTime int `json:"accessTime"`
+    AccessTime int64 `json:"accessTime"`
 }
